+++ src/components/forms/completions/CompletionQuantityTextField.tsx
@@ -1,12 +1,41 @@
 import styled from '@emotion/styled';
-import { TextField } from '@mui/material';
+import { Button, TextField, TextFieldProps } from '@mui/material';
+import React, { FC } from 'react';
 
-const CompletionQuantityTextField = styled(TextField)`
+const StyledButton = styled(Button)`
+  min-width: auto;
+`;
+const StyledTextField = styled(TextField)`
   fieldset {
     border-width: 0 1px 1px 0;
     border-top-right-radius: 0;
     border-bottom-right-radius: 0;
   }
 `;
+
+interface Props {
+  onDecrement?: () => void;
+  onIncrement?: () => void;
+  textFieldProps: TextFieldProps;
+}
+
+const CompletionQuantityTextField: FC<Props> = ({ onDecrement, onIncrement, textFieldProps }) => {
+  return (
+    <StyledTextField
+      inputProps={{
+        style: {
+          textAlign: 'center'
+        }
+      }}
+      InputProps={{
+        endAdornment: onIncrement && <StyledButton onClick={() => onIncrement()}>+</StyledButton>,
+        startAdornment: onDecrement && <StyledButton onClick={() => onDecrement()}>-</StyledButton>,
+        sx: { px: 1 }
+      }}
+      label={'Quantity'}
+      {...textFieldProps}
+    />
+  );
+};
 
 export default CompletionQuantityTextField;